# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller._ResponseCommunity:
    properties:
      cname:
        type: string
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      community_id:
        type: integer
      community_name:
        type: string
      creattime:
        type: string
      introduction:
        type: string
      message:
        description: 提示信息
        type: string
    type: object
  controller._ResponsePost:
    properties:
      author_id:
        type: integer
      author_name:
        type: string
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      community_id:
        type: integer
      community_name:
        type: string
      content:
        type: string
      creat_time:
        type: string
      creattime:
        type: string
      introduction:
        type: string
      message:
        description: 提示信息
        type: string
      post_approve_num:
        type: integer
      post_id:
        type: integer
      status:
        type: integer
      title:
        type: string
    required:
    - community_id
    - content
    - title
    type: object
  controller._ResponseUser:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      date:
        description: 返回数据
    type: object
  controller.ResCode:
    enum:
    - 10000
    - 10001
    - 10002
    - 10003
    - 10004
    - 10005
    - 10006
    - 10007
    type: integer
    x-enum-varnames:
    - SuccessCode
    - ErrorCodeInvalidParams
    - ErrorCodeUserAlreadyExist
    - ErrorCodeUserNotExist
    - ErrorCodeInvalidPassword
    - ErrorCodeServerBusy
    - ErrorCodInvalidAuth
    - ErrorCodeNeedLogin
  mod.ParamSignIn:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  mod.ParamSignUp:
    properties:
      password:
        type: string
      re_password:
        description: eqfield必须相同
        type: string
      username:
        type: string
    required:
    - password
    - re_password
    - username
    type: object
  mod.ParamsGetPostList:
    properties:
      community_id:
        description: 可以为空
        type: integer
      page:
        description: form 处理ShouldBindQuery参数的tag
        type: integer
      size:
        description: 每页数据量
        type: integer
      type:
        description: 排序依据
        example: score
        type: string
    type: object
  mod.Post:
    properties:
      author_id:
        type: integer
      community_id:
        type: integer
      content:
        type: string
      creat_time:
        type: string
      post_id:
        type: integer
      status:
        type: integer
      title:
        type: string
    required:
    - community_id
    - content
    - title
    type: object
  mod.PostVoted:
    properties:
      direction:
        description: "oneof 判断值在一个范围里\trequired如果传0，会判断为空"
        enum:
        - 1
        - -1
        - 0
        type: integer
      post_id:
        description: Userid从c中获取
        type: string
    required:
    - post_id
    type: object
host: localhost:8080
info:
  contact:
    email: yhelee99@gmail.com
    name: Yhelee
    url: https://github.com/Yhelee99
  description: 一个后端项目
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/Yhelee99
  title: Bluebell项目
  version: "1.0"
paths:
  /community:
    get:
      consumes:
      - application/json
      description: 获取社区列表接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunity'
      security:
      - ApiKeyAuth: []
      summary: 获取社区列表
      tags:
      - 社区相关接口
  /community/:id:
    get:
      consumes:
      - application/json
      description: 根据社区id查询社区详情
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 社区ID,可不传,不传获取全量
        format: int64
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseCommunity'
      security:
      - ApiKeyAuth: []
      summary: 获取社区列表
      tags:
      - 社区相关接口
  /createpost:
    post:
      consumes:
      - application/json
      description: 创建帖子接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 创建参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/mod.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 创建帖子
      tags:
      - 帖子相关接口
  /getpostslist:
    get:
      consumes:
      - application/json
      description: 可选排序方式的获取帖子列表的接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取帖子的参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/mod.ParamsGetPostList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePost'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子列表
      tags:
      - 帖子相关接口
  /post/:id:
    get:
      consumes:
      - application/json
      description: 根据postid获取帖子详情
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 帖子id
        format: int64
        in: path
        name: postid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePost'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子详情
      tags:
      - 帖子相关接口
  /post/voted:
    post:
      consumes:
      - application/json
      description: 帖子投票接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 获取帖子参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/mod.PostVoted'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 帖子投票
      tags:
      - 帖子相关接口
  /posts:
    get:
      consumes:
      - application/json
      description: 获取帖子列表接口
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页数
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponsePost'
      security:
      - ApiKeyAuth: []
      summary: 获取帖子列表
      tags:
      - 帖子相关接口
  /signin:
    post:
      consumes:
      - application/json
      description: 用于用户登录
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/mod.ParamSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseUser'
      security:
      - ApiKeyAuth: []
      summary: 用户登录接口
      tags:
      - 用户相关
  /signup:
    post:
      consumes:
      - application/json
      description: 用于用户注册
      parameters:
      - description: 用户信息
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/mod.ParamSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: 用户注册接口
      tags:
      - 用户相关
swagger: "2.0"